somewhat major refactor
pulled config stuff into a json file
this allows me to read the config once and pass it around to plugins
for plugins using the mentionName, this is necessary.

also pulling some plugins out of the main file and into the plugins dir
this is not the final way this will work but I get nervous so small
steps.